pyproject.toml
[tool.poetry]
authors = ["Robert Kaussow <mail@thegeeklab.de>"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "Natural Language :: English",
  "Operating System :: POSIX",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Topic :: Utilities",
  "Topic :: Software Development",
  "Topic :: Software Development :: Documentation",
]
description = "Generate documentation from annotated Ansible roles using templates."
documentation = "https://ansible-doctor.geekdocs.de/"
homepage = "https://ansible-doctor.geekdocs.de/"
include = [
  "LICENSE",
]
keywords = ["ansible", "code", "review"]
license = "GPL-3.0-only"
name = "ansible-doctor"
packages = [
  {include = "ansibledoctor"},
]
readme = "README.md"
repository = "https://github.com/thegeeklab/ansible-doctor/"
version = "0.0.0"

[tool.poetry.dependencies]
Jinja2 = "^2.11.2"
anyconfig = "^0.9.11"
appdirs = "^1.4.4"
colorama = "^0.4.4"
environs = "^9.3.0"
jsonschema = "^3.2.0"
nested-lookup = "^0.2.21"
pathspec = "^0.8.1"
python = "^3.6.0"
python-json-logger = "^2.0.1"
"ruamel.yaml" = "^0.16.12"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
flake8 = "^3.8.4"
flake8-blind-except = "^0.1.1"
flake8-builtins = "^1.5.3"
flake8-colors = "^0.1.9"
flake8-docstrings = "^1.5.0"
flake8-eradicate = "^1.0.0"
flake8-isort = "^4.0.0"
flake8-logging-format = "^0.6.0"
flake8-pep3101 = "^1.3.0"
flake8-polyfill = "^1.0.2"
flake8-quotes = "^3.2.0"
pep8-naming = "^0.11.1"
pydocstyle = "^5.1.1"
pytest = "^6.2.1"
pytest-cov = "^2.10.1"
pytest-mock = "^3.4.0"
yapf = "^0.30.0"

[tool.poetry.scripts]
ansible-doctor = "ansibledoctor.cli:main"

[tool.poetry-dynamic-versioning]
enable = true
style = "semver"
vcs = "git"

[tool.isort]
default_section = "THIRDPARTY"
force_single_line = true
line_length = 99
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip_glob = ["**/.env*", "**/env/*", "**/docs/*"]

[tool.pytest.ini_options]
addopts = "--cov=ansibledoctor --cov-report xml:coverage.xml --cov-append --no-cov-on-fail"
filterwarnings = [
  "ignore::FutureWarning",
  "ignore:.*collections.*:DeprecationWarning",
  "ignore:.*pep8.*:FutureWarning",
]

[tool.coverage.run]
omit = ["**/test/*"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
